requirements.txt
attrs==18.2.0
biopython==1.73
click==7.0
lxml==4.2.1
ratelimiter==1.2.0.post0
requests==2.20.0
retry==0.9.2
